+++ examples/angular-reactive-forms/src/app/reactive-form/reactive-form.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormControl, FormGroup } from '@angular/forms';
+import { UntypedFormControl, UntypedFormGroup } from '@angular/forms';
 
 @Component({
   selector: 'app-reactive-form',
@@ -13,10 +13,10 @@ export class ReactiveFormComponent implements OnInit {
   ngOnInit(): void {
   }
 
-  signupForm = new FormGroup({
-    username: new FormControl('admin'),
-    password: new FormControl({ value: '', disabled: false }),
-    consent: new FormControl()
+  signupForm = new UntypedFormGroup({
+    username: new UntypedFormControl('admin'),
+    password: new UntypedFormControl({ value: '', disabled: false }),
+    consent: new UntypedFormControl()
   });
 
   onSubmit() {