+++ examples/nuxt-ssr/nuxt.config.js
@@ -1,5 +1,4 @@
 export default {
-  mode: 'universal',
   /*
    ** Headers of the page
    */