+++ packages/components/src/components/dropdown-select/dropdown-select.tsx
@@ -13,7 +13,7 @@ import {
 import classNames from 'classnames';
 import statusNote from '../../utils/status-note';
 import { computePosition } from '@floating-ui/dom';
-import { emitEvent } from '../../utils/utils';
+import { emitEvent, generateUniqueId } from '../../utils/utils';
 
 enum Actions {
   Close = 'Close',
@@ -184,6 +184,7 @@ export class DropdownSelect {
 
   @Prop() comboboxId?: string = 'combobox';
   @Prop() label: string;
+  @Prop() name?: string;
   @Prop() helperText?: string = '';
   @Prop() disabled?: boolean;
   @Prop() readonly?: boolean;
@@ -381,7 +382,7 @@ export class DropdownSelect {
       readOptions(this.hostElement).find(({ value }) => value === this.value) ||
       ({} as any)
     ).ItemElement;
-    const helperTextId = `helper-message`;
+    const helperTextId = `helper-message-${generateUniqueId()}`;
     const ariaDescribedByAttr = { 'aria-describedBy': helperTextId };
 
     return (