+++ packages/components/src/components/progress-bar/progress-bar.tsx
@@ -85,15 +85,12 @@ export class ProgressBar {
   `;
 
   progressStyle = () => {
+    const customColor = this.customColor
+      ? { '--background': this.customColor }
+      : {};
     return {
-      width: this.disabled ? '100%' : `${this.percentage}%`,
-      border: '1px solid transparent',
-      background: this.customColor
-        ? this.customColor
-        : this.disabled
-        ? 'var(--background-disabled)'
-        : `var(--background)`,
-      animation: this.disabled ? 'none' : 'showProgress 3s ease-in-out',
+      '--progress': this.disabled ? '100%' : `${this.percentage}%`,
+      ...customColor,
     };
   };
 
@@ -151,11 +148,13 @@ export class ProgressBar {
               aria-label={this.label}
               id={this.progressBarId}
             >
-              <div
-                part="inner"
-                class="progress-bar__inner"
-                style={this.progressStyle()}
-              ></div>
+              {this.percentage > 0 && (
+                <div
+                  part="inner"
+                  class="progress-bar__inner"
+                  style={this.progressStyle()}
+                ></div>
+              )}
             </div>
             <slot name="icon"></slot>
           </div>