+++ packages/components/src/components/tab-nav/tab-nav.tsx
@@ -35,24 +35,43 @@ export class TabNav {
   @Prop() small?: boolean = false;
   /** (optional) size  */
   @Prop() size: 'small' | 'large' = 'small';
+  /** (optional) autoFocus  */
+  @Prop() autoFocus: boolean = false;
   /** (optional) Injected CSS styles */
   @Prop() styles?: string;
 
   @Listen('click')
   handleClick(event: MouseEvent) {
-    this.removeFirstRenderAttr();
+    // To provent event bubbling.
+    event.stopPropagation();
 
-    const nextTab = event.target as HTMLScaleTabHeaderElement;
-    if (nextTab.getAttribute('role') !== 'tab') {
-      return;
+    // workaround for slotted icons
+    const targetHTMLElement = event.target as HTMLElement;
+    const targetTag = targetHTMLElement.tagName.toLowerCase();
+    const svgTags = ['svg', 'g', 'path'];
+    let nextTab: HTMLScaleTabHeaderElement;
+
+    if (svgTags.includes(targetTag)) {
+      const closestNextTab = targetHTMLElement.closest(
+        `scale-tab-header[role="tab"]`
+      ) as HTMLScaleTabHeaderElement;
+      if (closestNextTab) {
+        nextTab = closestNextTab;
+        this.selectTab(nextTab);
+      }
+    } else {
+      if (
+        (event.target as HTMLScaleTabHeaderElement).getAttribute('role') ===
+        'tab'
+      ) {
+        nextTab = event.target as HTMLScaleTabHeaderElement;
+        this.selectTab(nextTab);
+      }
     }
-    this.selectTab(nextTab);
   }
 
   @Listen('keydown')
   handleKeydown(event: KeyboardEvent) {
-    this.removeFirstRenderAttr();
-
     const target = event.target as HTMLScaleTabHeaderElement;
     let nextTab;
 
@@ -83,13 +102,6 @@ export class TabNav {
 
     event.preventDefault();
     this.selectTab(nextTab);
-  }
-
-  removeFirstRenderAttr() {
-    const tabs = this.getAllEnabledTabs();
-    tabs.forEach((tab) => {
-      tab.removeAttribute('first-render');
-    });
   }
 
   connectedCallback() {
@@ -155,22 +167,15 @@ export class TabNav {
 
   linkPanels() {
     const tabs = this.getAllEnabledTabs();
-    const selectedTab = tabs.find((x) => x.selected);
+    const selectedTab = tabs.find((x) => x.selected) || tabs[0];
 
     tabs.forEach((tab) => {
       const panel = tab.nextElementSibling;
       tab.setAttribute('aria-controls', panel.id);
+      tab.setAttribute('auto-focus', this.autoFocus.toString());
       panel.setAttribute('aria-labelledby', tab.id);
-      if (!selectedTab) {
-        // we pass this down to tab-header to prevent the first element to be focused on first render (a11y)
-        tab.setAttribute('first-render', 'true');
-      }
     });
-    if (selectedTab) {
-      this.selectTab(selectedTab);
-    } else {
-      this.selectTab(tabs[0]);
-    }
+    this.selectTab(selectedTab);
   }
 
   reset() {