+++ packages/components/src/components/telekom/telekom-header/telekom-header.tsx
@@ -29,10 +29,10 @@ import cx from 'classnames';
 export class TelekomHeader {
   @Element() hostElement: HTMLStencilElement;
 
-  @Prop() mainNavigation: any;
-  @Prop() appName?: string;
+  @Prop({ reflect: true }) appName?: string;
   @Prop() appNameLink?: string;
   @Prop() appNameClick?: any;
+  @Prop() type: string = 'default';
 
   @State() scrolled: boolean;
   @State() scrolledBack: boolean = false;
@@ -48,63 +48,67 @@ export class TelekomHeader {
 
   render() {
     return (
-      <Host>
+      <Host
+        scrolled={this.type !== 'subtle' && this.scrolled}
+        scrolled-back={this.type !== 'subtle' && this.scrolledBack}
+      >
         <header
-          part={cx('base', {
-            scrolled: this.scrolled,
-            'scrolled-back': this.scrolledBack,
+          part={cx('base', this.type, {
+            scrolled: this.type !== 'subtle' && this.scrolled,
+            'scrolled-back': this.type !== 'subtle' && this.scrolledBack,
           })}
         >
-          <div part="container">
-            <slot name="app-logo">
-              <scale-logo part="app-logo" variant="white"></scale-logo>
-            </slot>
+          <div part="fixed-wrapper">
+            <div part="container">
+              <slot name="logo">
+                <scale-logo part="app-logo" variant="white"></scale-logo>
+              </slot>
 
-            <div part="horizontal-menus">
-              <div part="extended-menu">
-                <div part="extended-menu-left">
+              <div part="body">
+                <div part="top-bar">
                   {this.appName ? (
-                    <div part="app-name-extended">
-                      <scale-telekom-app-name>
-                        {this.appNameLink ? (
-                          <a
-                            onClick={this.appNameClick}
-                            href={this.appNameLink}
-                          >
-                            {this.appName}
-                          </a>
-                        ) : (
-                          <span>{this.appName}</span>
-                        )}
-                      </scale-telekom-app-name>
-                    </div>
-                  ) : null}
-
-                  <slot name="meta-nav-ext"></slot>
-                </div>
-                <div part="extended-menu-right">
-                  <slot name="meta-nav"></slot>
-                  <slot name="language-switch"></slot>
-                </div>
-              </div>
-
-              <div part="app-name-and-base-menu">
-                {this.appName ? (
-                  <div part="app-name">
-                    <scale-telekom-app-name>
+                    <div part="top-app-name">
                       {this.appNameLink ? (
-                        <a onClick={this.appNameClick} href={this.appNameLink}>
+                        <a
+                          part="app-name-text"
+                          onClick={this.appNameClick}
+                          href={this.appNameLink}
+                        >
                           {this.appName}
                         </a>
                       ) : (
-                        <span>{this.appName}</span>
+                        <span part="app-name-text">{this.appName}</span>
                       )}
-                    </scale-telekom-app-name>
+                    </div>
+                  ) : null}
+
+                  <div part="top-body">
+                    <slot name="meta-nav-external"></slot>
+                    <slot name="meta-nav"></slot>
+                    <slot name="lang-switcher"></slot>
                   </div>
-                ) : null}
-                <div part="base-menu">
-                  <slot name="main-nav"></slot>
-                  <slot name="functions"></slot>
+                </div>
+
+                <div part="bottom-bar">
+                  {this.appName ? (
+                    <div part="bottom-app-name">
+                      {this.appNameLink ? (
+                        <a
+                          part="app-name-text"
+                          onClick={this.appNameClick}
+                          href={this.appNameLink}
+                        >
+                          {this.appName}
+                        </a>
+                      ) : (
+                        <span part="app-name-text">{this.appName}</span>
+                      )}
+                    </div>
+                  ) : null}
+                  <div part="bottom-body">
+                    <slot name="main-nav"></slot>
+                    <slot name="functions"></slot>
+                  </div>
                 </div>
               </div>
             </div>