+++ packages/storybook-vue/stories/components/text-field/ScaleTextField.vue
@@ -18,10 +18,12 @@
     :counter="counter"
     :size="size"
     :value="value"
+    :aria-detailed-id="ariaDetailedId"
     :input-id="inputId"
     :input-autofocus="inputAutofocus"
     :input-autocomplete="inputAutocomplete"
     :experimental-controlled="experimentalControlled"
+    :hide-label-visually="hideLabelVisually"
     @scaleInput="scaleInput"
     @scaleChange="scaleChange"
     @scaleFocus="scaleFocus"
@@ -55,10 +57,12 @@ export default {
     counter: { type: Boolean },
     size: { type: String },
     value: { type: String },
+    ariaDetailedId: { type: String },
     inputId: { type: String },
     inputAutofocus: { type: Boolean },
     inputAutocomplete: { type: String },
     experimentalControlled: { type: Boolean },
+    hideLabelVisually: {type: Boolean, default: false},
     styles: { type: String },
   },
   methods: {